requirements/base.txt
pyyaml==3.11
tornado==4.1
pymysql==0.6.4
mongoengine==0.8.7
pymongo==2.8
--allow-external sysv_ipc
--allow-unverified sysv_ipc 
sysv_ipc==0.6.8
boto==2.36.0
python-gcm==0.1.5
apns==2.0.1
requests==2.5.3
